[refactor] Moved file validation logic to another file (#23)
